fix: 락 시간 타입 수정

+++ src/main/java/com/example/travelshooting/common/Const.java
@@ -27,8 +27,8 @@ public class Const {
     // 동시성 제어
     public static final String REDISSON_PREFIX = "redis://";
     public static final String RESERVATION_LOCK_PREFIX = "lock:reservation:";
-    public static final int RESERVATION_LOCK_WAIT_TIME = 5;
-    public static final int RESERVATION_LOCK_LEASE_TIME = 10;
+    public static final Long RESERVATION_LOCK_WAIT_TIME = 5L;
+    public static final Long RESERVATION_LOCK_LEASE_TIME = 10L;
 
     // 캐싱
     public static final String COMPANY_CACHE_PREFIX = "companies:page:";
